backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg[asyncio]==3.1.13
python-multipart==0.0.6
PyMuPDF==1.23.8
python-docx==0.8.11
spacy==3.7.2
sentence-transformers==2.2.2
pylatex==1.4.2
openai-whisper==20231117
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.1.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
fastapi-cors==0.0.6
